initialize the resources list inside the cluster loop
This commit moves the resources list initialization inside the cluster
loop. This is required to start with an empty list for each cluster
instead of a list already filed with the resources of the previous cluster.

Fix a bug where each cluster is listing resources that are only existing
in the cluster previously listed.